Rename test_utils to utils so that it is not considered a test module by pytest.
