+++ genshi/filters/tests/test_html.py
@@ -18,7 +18,7 @@ import six
 from genshi.input import HTML, ParseError
 from genshi.filters.html import HTMLFormFiller, HTMLSanitizer
 from genshi.template import MarkupTemplate
-from genshi.tests.test_utils import doctest_suite
+from genshi.tests.utils import doctest_suite
 
 class HTMLFormFillerTestCase(unittest.TestCase):
