+++ genshi/tests/utils.py
